Small typo
Was reading the docs and saw something small.

I did try to click the Go library documentation thinking it was the docs on how to use this in a normal Go app, but was surprised it was a link to the Go.dev docs. Any plans, I know it’s early, to document integration with a Go app?