Add Joomla on FrankenPHP Example
Hi, FrankenPHP Team,
This pull request is a suggestion to add Joomla on FrankenPHP as it works too. I tried it and made a docker image for it based on your frankenphp-worpress repo. If you don't mind, please add the Joomla version too on your README.md as an example, besides Wordpress and Drupal.