Add documentation for 404 default
When a route doesn't exist, caddy sets a 200 response code. However, PHP response emitters (such as Nyholm's) do NOT set a 404 when a route doesn't exist because that is the default for other web servers.

This hints that a default can be set prior to running the application if users see a 200 when they expect a 404.